+++ EditorConfiguration.c
@@ -6,11 +6,14 @@
  * Configuration of global options of PKS edit. These are for now 
  * saved in a file PKSEDIT.INI by default located in the PKS_SYS folder.
  *
- * 										created: 
- * 										last modified:
- *										author: Tom
- *
- * (c) Pahlen & Krauss
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ * 
+ * created: 
+ * last modified:
+ * author: Tom
+ * 
  */
 
 #include <windows.h>